+++ src/main/java/com/nu/art/cyborg/core/CyborgControllerBase.java
@@ -467,7 +467,7 @@ abstract class CyborgControllerBase
 	}
 
 	@Override
-	public final void postOnUI(int delay, Runnable action) {
+	public final void postOnUI(long delay, Runnable action) {
 		cyborg.postOnUI(delay, action);
 	}
 
@@ -477,7 +477,7 @@ abstract class CyborgControllerBase
 	}
 
 	@Override
-	public final void removeAndPostOnUI(int delay, Runnable action) {
+	public final void removeAndPostOnUI(long delay, Runnable action) {
 		cyborg.removeAndPostOnUI(delay, action);
 	}
