+++ src/main/java/com/nu/art/cyborg/core/CyborgModule.java
@@ -282,7 +282,7 @@ public abstract class CyborgModule
 	 * Interfaces .....
 	 */
 	@Override
-	public final void postOnUI(int delay, Runnable action) {
+	public final void postOnUI(long delay, Runnable action) {
 		cyborg.postOnUI(delay, action);
 	}
 
@@ -292,7 +292,7 @@ public abstract class CyborgModule
 	}
 
 	@Override
-	public final void removeAndPostOnUI(int delay, Runnable action) {
+	public final void removeAndPostOnUI(long delay, Runnable action) {
 		cyborg.removeAndPostOnUI(delay, action);
 	}
